+++ 502_Calculator_WS_Client/src/calcws/AddResponse.java
@@ -1,6 +1,8 @@
 
 package calcws;
 
+import java.math.BigInteger;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
@@ -17,7 +19,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="return" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="return" type="{http://www.w3.org/2001/XMLSchema}BigInteger"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -33,13 +35,13 @@ import javax.xml.bind.annotation.XmlType;
 public class AddResponse {
 
     @XmlElement(name = "return")
-    protected int _return;
+    protected BigInteger _return;
 
     /**
      * Gets the value of the return property.
      * 
      */
-    public int getReturn() {
+    public BigInteger getReturn() {
         return _return;
     }
 
@@ -47,7 +49,7 @@ public class AddResponse {
      * Sets the value of the return property.
      * 
      */
-    public void setReturn(int value) {
+    public void setReturn(BigInteger value) {
         this._return = value;
     }
